+++ warp/native/volume.h
@@ -6,7 +6,7 @@
 #include "builtin.h"
 
 #include "array.h"
-#include "mat.h"
+#include "mat_ops.h"
 
 #define PNANOVDB_C
 #define PNANOVDB_MEMCPY_CUSTOM